pyproject.toml
[tool.poetry]
name = "python-raft"
version = "0.1.0"
description = ""
authors = ["zghp <fatmaivonakoc10@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "3.9.3"
aiosignal = "1.3.1"
annotated-types = "0.6.0"
anyio = "4.3.0"
async-timeout = "4.0.3"
asyncclick = "8.1.7.2"
asyncio = "3.4.3"
attrs = "23.2.0"
bcrypt = "4.1.2"
beautifulsoup4 = "4.12.3"
boto3 = "1.34.70"
botocore = "1.34.70"
cdrouter = "0.9.7"
certifi = "2024.2.2"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
cryptography = "42.0.5"
decorator = "5.1.1"
deprecated = "1.2.14"
fabric = "3.2.2"
frozenlist = "1.4.1"
future = "1.0.0"
h11 = "0.14.0"
idna = "3.6"
invoke = "2.2.0"
jmespath = "1.0.1"
marshmallow = "3.21.1"
multidict = "6.0.5"
numpy = "1.26.4"
opencv-python = "4.9.0.80"
outcome = "1.3.0.post0"
packaging = "24.0"
paramiko = "3.4.0"
pillow = "10.2.0"
pipdeptree = "2.16.2"
pycparser = "2.21"
pydantic = "2.6.4"
pydantic-core = "2.16.3"
pynacl = "1.5.0"
pyserial = "3.5"
pysocks = "1.7.1"
pytesseract = "0.3.10"
python-dateutil = "2.9.0.post0"
python-kasa = "0.6.2.1"
pyyaml = "6.0.1"
requests = "2.31.0"
requests-toolbelt = "1.0.0"
s3transfer = "0.10.1"
selenium = "4.18.1"
six = "1.16.0"
sniffio = "1.3.1"
sortedcontainers = "2.4.0"
soupsieve = "2.5"
trio = "0.25.0"
trio-websocket = "0.11.1"
typing-extensions = "4.10.0"
urllib3 = "1.26.18"
wrapt = "1.16.0"
wsproto = "1.2.0"
yarl = "1.9.4"
pytest = "^8.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
